requirements.txt
pandas==1.3.3
openai==0.27.0
python-dotenv==0.19.1